change files and change type_statement -> tupe_structureand old type_structure is remove
