+++ lang-common/src/type_defines/additional_defines/additional_statement.rs
@@ -1,0 +1,21 @@
+use crate::types::{property_key::PropertyKey, type_name::TypeName};
+
+pub trait AdditionalStatement {
+    fn get_type_comment(&self, type_name: &TypeName) -> Option<String>;
+    fn get_property_comment(
+        &self,
+        type_name: &TypeName,
+        property_key: &PropertyKey,
+    ) -> Option<String>;
+    fn get_type_attribute(
+        &self,
+        type_name: &TypeName,
+        property_key: &PropertyKey,
+    ) -> Option<String>;
+    fn get_property_attribute(
+        &self,
+        type_name: &TypeName,
+        property_key: &PropertyKey,
+    ) -> Option<String>;
+    fn is_property_optional(&self, type_name: &TypeName, property_key: &PropertyKey) -> bool;
+}
