+++ lang-common/src/type_defines/additional_defines/attribute_store.rs
@@ -2,29 +2,29 @@ use std::collections::HashMap;
 
 use utils::store_fn::push_to_kv_vec;
 
-use super::structures::{PropertyKey, TypeName};
+use crate::types::{property_key::PropertyKey, type_name::TypeName};
 
-pub type Comment = String;
-pub struct CommentStore<'a> {
-    type_store: HashMap<&'a TypeName, Vec<Comment>>,
-    property_store: HashMap<(&'a TypeName, &'a PropertyKey), Vec<Comment>>,
+pub type Attribute = String;
+pub struct AttributeStore<'a> {
+    type_store: HashMap<&'a TypeName, Vec<Attribute>>,
+    property_store: HashMap<(&'a TypeName, &'a PropertyKey), Vec<Attribute>>,
 }
 
-impl<'a> CommentStore<'a> {
+impl<'a> AttributeStore<'a> {
     pub fn new() -> Self {
         Self {
             type_store: HashMap::new(),
             property_store: HashMap::new(),
         }
     }
-    pub fn add_type_comment(&mut self, type_name: &'a TypeName, attribute: Comment) {
+    pub fn add_type_attribute(&mut self, type_name: &'a TypeName, attribute: Attribute) {
         push_to_kv_vec(&mut self.type_store, type_name, attribute);
     }
-    pub fn add_property_comment(
+    pub fn add_property_attribute(
         &mut self,
         type_name: &'a TypeName,
         property_key: &'a PropertyKey,
-        attribute: Comment,
+        attribute: Attribute,
     ) {
         push_to_kv_vec(
             &mut self.property_store,
@@ -32,14 +32,14 @@ impl<'a> CommentStore<'a> {
             attribute,
         );
     }
-    pub fn get_type_comment(&self, type_name: &TypeName) -> Option<&Vec<Comment>> {
+    pub fn get_type_attribute(&self, type_name: &TypeName) -> Option<&Vec<Attribute>> {
         self.type_store.get(type_name)
     }
-    pub fn get_property_comment(
+    pub fn get_property_attribute(
         &self,
         type_name: &'a TypeName,
         property_key: &'a PropertyKey,
-    ) -> Option<&Vec<Comment>> {
+    ) -> Option<&Vec<Attribute>> {
         self.property_store.get(&(type_name, property_key))
     }
 }
