+++ lang-common/src/types.rs
@@ -1,6 +1,5 @@
-pub(crate) mod attribute_store;
-pub(crate) mod comment_store;
-pub mod froms;
-pub(crate) mod optional_type_store;
-pub mod statement;
+pub mod primitive_type;
+pub mod property_key;
+pub mod property_type;
 pub mod structures;
+pub mod type_name;
