--- lang-common/src/types/froms.rs
@@ -1,1 +1,0 @@
-pub mod from_json;
